Add all classes into duke package
