+++ src/main/java/duke/exceptions/EmptyDescriptionException.java
@@ -1,7 +1,9 @@
+package duke.exceptions;
+
 public class EmptyDescriptionException extends  Exception {
 
     public EmptyDescriptionException() {
-        super("bro where's the task description?");
+        super("bro where's the duke.task description?");
     }
 
     public EmptyDescriptionException(String message) {
