+++ src/main/java/duke/exceptions/InvalidTaskException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class InvalidTaskException extends Exception {
 
     public InvalidTaskException() {
