+++ src/main/java/duke/storage/Storage.java
@@ -1,0 +1,5 @@
+package duke.storage;
+
+public class Storage {
+    
+}
