+++ src/main/java/duke/task/Deadline.java
@@ -1,6 +1,7 @@
+package duke.task;
+
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
+
 public class Deadline extends Task {
 
     protected LocalDate date;
