+++ src/main/java/duke/task/Event.java
@@ -1,6 +1,7 @@
+package duke.task;
+
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
+
 public class Event extends Task {
 
     protected LocalDate from;
