+++ src/main/java/duke/ui/Ui.java
@@ -1,0 +1,15 @@
+package duke.ui;
+
+import duke.task.TaskList;
+
+public class Ui {
+
+    public Ui(TaskList taskList) {
+    }
+
+    public void greet() {
+        System.out.println("Hello! I'm Mash \nWhat can I do for you?");
+    }
+
+
+}
