+++ text-ui-test/EXPECTED.TXT
@@ -1,24 +1,24 @@
 Hello! I'm Mash 
 What can I do for you?
-Got it. I've added this task: 
+Got it. I've added this duke.task:
 [T][ ] eat
 Now you have 1 tasks in the list.
-Got it. I've added this task: 
+Got it. I've added this duke.task:
 [D][ ] play (by: 5pm)
 Now you have 2 tasks in the list.
-Got it. I've added this task: 
+Got it. I've added this duke.task:
 [E][ ] sleep (from: 12am to: 6am)
 Now you have 3 tasks in the list.
-Nice! I've marked this task as done:
+Nice! I've marked this duke.task as done:
 [D][X] play (by: 5pm)
-OK, I've marked this task as not done yet:
+OK, I've marked this duke.task as not done yet:
 [D][ ] play (by: 5pm)
 1. [T][ ] eat
 2. [D][ ] play (by: 5pm)
 3. [E][ ] sleep (from: 12am to: 6am)
-Noted. I've removed this task:
+Noted. I've removed this duke.task:
 [E][ ] sleep (from: 12am to: 6am)
 Now you have 2 tasks in the list.
 1. [T][ ] eat
 2. [D][ ] play (by: 5pm)
-☹ OOPS!!! The description of a task cannot be empty.
+☹ OOPS!!! The description of a duke.task cannot be empty.
